Biblioteca SweetAlert

+++ webapp/assets/js/publicacoes.js
@@ -19,7 +19,7 @@ function criarPublicacao(evento) {
     }).done(function () {
         window.location = "/home";
     }).fail(function () {
-        console.log("Erro ao criar publicação.");
+        Swal.fire("Ops...", "Erro ao criar a publicação!", "error");
     })
 }
 
@@ -44,7 +44,7 @@ function curtirPublicacao(evento) {
         elementoClicado.removeClass("curtir-publicacao");
 
     }).fail(function () {
-        console.log("Erro ao curtir a publicação!");
+        Swal.fire("Ops...", "Erro ao curtir a publicação!", "error");
     }).always(function () {
         elementoClicado.prop("disabled", false);
     });
@@ -71,7 +71,7 @@ function descurtirPublicacao(evento) {
         elementoClicado.addClass("curtir-publicacao");
 
     }).fail(function () {
-        console.log("Erro ao curtir a publicação!");
+        Swal.fire("Ops...", "Erro ao descurtir a publicação!", "error");
     }).always(function () {
         elementoClicado.prop("disabled", false);
     });
@@ -89,11 +89,14 @@ function atualizarPublicacao(evento) {
             titulo: $("#titulo").val(),
             conteudo: $("#conteudo").val()
         }
-    }).done(function() {
-        alert("Publicação atualizada com sucesso!");
-    }).fail(function() {
-        alert("Erro ao atualizar a publicação.");
-    }).always(function() {
+    }).done(function () {
+        Swal.fire("Sucesso!", "Publicação editada com sucesso!", "success")
+            .then(function () {
+                window.location = "/home";
+            })
+    }).fail(function () {
+        Swal.fire("Ops...", "Erro ao editar a publicação!", "error");
+    }).always(function () {
         $("#atualizar-publicacao").prop("disabled", false);
     })
 }
@@ -101,20 +104,31 @@ function atualizarPublicacao(evento) {
 function deletarPublicacao(evento) {
     evento.preventDefault();
 
-    const elementoClicado = $(evento.target);
-    const publicacao = elementoClicado.closest("div");
-    const publicacaoId = publicacao.data("publicacao-id");
+    Swal.fire({
+        title: 'Atenção!',
+        text: "Tem certeza que deseja deletar essa publicação? Essa ação é irreversível!",
+        showCancelButton: true,
+        cancelButtonText: "Cancelar",
+        icon: 'warning'
+    }).then(function (confirmacao) {
+        if (!confirmacao.value) return;
 
-    elementoClicado.prop("disabled", true);
+        const elementoClicado = $(evento.target);
+        const publicacao = elementoClicado.closest("div");
+        const publicacaoId = publicacao.data("publicacao-id");
 
-    $.ajax({
-        url: `/publicacoes/${publicacaoId}`,
-        method: "DELETE"
-    }).done(function() {
-        publicacao.fadeOut("slow", function() {
-            $(this).remove();
-        })
-    }).fail(function() {
-        alert("Erro ao deletar a publicação.");
-    });
+        elementoClicado.prop("disabled", true);
+
+        $.ajax({
+            url: `/publicacoes/${publicacaoId}`,
+            method: "DELETE"
+        }).done(function () {
+            publicacao.fadeOut("slow", function () {
+                $(this).remove();
+            })
+        }).fail(function () {
+            Swal.fire("Ops...", "Erro ao deletar a publicação!", "error");
+        });
+    })
+
 }
